Implement cleaner Prim's algorithm
The primary difference between this commit and the
previous one is that this commit avoids creating
a path cell whose diagonal touches another path cell.
This singular characteristic was not necessarily
difficult to imagine conceptually--but implementing
such a characteristic was very tedious and confusing.

Essentially, when a wall is 'likely' to become a path
cell, we need to keep track of what cell is attached
to the wall that may become a path cell. By 'likely',
I mean that there is only one visited neighboring cell
touching the wall (i.e. the path cell it came from).
This visited neighboring cell is only from the top,
right, bottom, or left of the wall. Now, we need to
check the two diagonal cells in the opposite direction
of the attached cell.

Example: If a wall being checked came from a left cell,
we need to check the diagonals in the right direction.
These diags would be the top right and bottom right.
If they've been visited, do not make the path cell
and pop the wall from the wall list.